feat: capture authorization error in fetchwithvalidator
